Allow narrow unsigned operands on bpsMul/mulDiv and bump 0.1.3.
bps may be u8–u64 and mulDiv divisor c may be narrower than a/b, with
on-chain promote; ship SDK 0.1.3 aligned to ifx/ifx-core 0.1.2 (redeploy).

+++ rust-sdk/tests/common/planners/personal_amm.rs
@@ -108,7 +108,7 @@ pub fn plan_personal_amm_swap_instructions(
     } else {
         batch.let_eval(expr::bps_mul_floor(
             expr::r(&amount_out_gross_u64),
-            expr::u64(BPS_DENOM - fee_bps),
+            expr::u16((BPS_DENOM - fee_bps) as u16),
         ))?
     };
     let min_out = batch.let_const_u64(params.min_out)?;
